chore: Change convert signature in all other resources (#3940)
Continuation of #3926.

It's the last PR aiming to improve SDK's stability and reliability by
including an error in the return from conversion attempt. Till now, the
errors there were only logged which caused unexpected behavior in some
cases (like improper id parsing for functions and , because of that, not
finding the function with an empty id).

The change contains:

- Migrate all other objects that are not generated from defs
- Alter generation
- Migrate all other objects that are generated
- Rename convertRowsErr to convertRows when all objects are migrated
- Rename convertErr to convert when all objects are migrated
- Remove convertibleRowDeprecated when all objects are migrated
- Add compile checks like var _ convertibleRowDeprecated[DatabaseRole] =
new(databaseRoleDBRow) for each convertible (generated)

Future improvements:
- generated conversion tests (or at least sanity checks)
- the generated guards are not currently wrapped with other vars during
the generation and there are whitespace issues