chore: Fix check destroy logic (#3967)
CheckDestroy was silently failing, as we expected an error from
Snowflake when trying to fetch the objects as part of CheckDestroy. The
problem was that using the old ID parsing logic, we were always getting
AccountObjectIdentifier, and later casting was returning an error, but
it was treated as a successful check for an object not existing in
Snowflake.
Changes:
- checking casting error before succeeding
- logic split into legacy parsing and new id parsing
- cosmetic changes
- fixed listing and storage integration