chore: Add resource ID generation to migration script (#3993)
## Changes
- Add resource ID generation
- The generation was based on the properties used for resource grouping,
with a few exceptions (privileges were not taken into account, and the
account name). We don't need privileges in IDs for them to be unique
within one script run (we are grouping similar grant resources to have
all related privileges in the same resource), and the account name
doesn't need to be included, as grants can only be granted on the
current account (this I'm not 100% sure of as there may be some minor
cases where this may occur, but afaik only current account can
participate in grants to account role, so for majority, or all cases,
the account name will be the same).
- The dot in resource ID mapping is only added for visual clarity with
multi-part identifiers
- Tested with Terraform that even 500-character resource names are
handled, so we should be good regarding the resource name length

## Next prs
Document the limitations of generated resource names