fix: Fix privileges validation in `grant_privileges_to_account_role` resource (#3994)
Fix privileges validation in `grant_privileges_to_account_role`
resource. This bug was introduced in
https://github.com/snowflakedb/terraform-provider-snowflake/pull/3953.
According to [HC
docs](https://developer.hashicorp.com/terraform/language/expressions/references#values-not-yet-known):
> For example, adding a known value to an unknown value automatically
produces an unknown value as a result.

However, in this case, the value is not marked as unknown during the
plan (also shown in the PreApply assertion). During the validation,
however, the whole `privilege` field is also NOT unknown, but the item
type IS unknown, which caused panic.

## References
#3992 

## TODO
Add an SDKv2 functional test for this behavior.