feat: add semantic view resource - 1 (#4029)
Add a basic semantic view resource.
- Add a schema for the required fields.
- Do not enable the resource in the prod provider yet, as we expect a
few more PRs before this resource is usable, and we don't want to block
the releases.
Do not enable the resource in the test provider yet, as it can't be
created in its current form. It requires either the metrics or
dimensions block, which will be implemented next.
- Fix a typo in the Makefile.

## TODO
- Add the other supported fields & add this resource to the test
provider.
- Handle external changes.
- Handle object renaming.
- Handle copy grants.
- Write more acceptance tests.
- Provide examples.
- Add an entry to the migration guide.
- Add a data source.
- Extract common code, especially for block handling.

---------

Co-authored-by: shreyas <sarur@snowflake.com>
Co-authored-by: Jakub Michalak <jakub.michalak@snowflake.com>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>