chore: Adjust database acceptance tests to set the common testing pattern (#4113)
## Changes
Adjust database tests to set an example on a common testing pattern for
main Terraform use cases for a given object. This PR only includes the
testing pattern for resources that represent an object on the Snowflake
side. See the next PRs section on checking other tests, as they may
differ from the ones set here. By basing on a few acceptance tests
already within the project, I ended up with

~~two tests (BasicUseCase and CompleteUseCase; the naming may be
confusing because it refers more to the initial configuration rather
than overall complexity, but the naming can be adjusted later on). Both
are pretty minimal, but they validate basic CRUD + import + external
changes functionality of a resource, which should be enough for our use
case, and other tests should be used to fully test the resource's
functionality.~~

**Update:** one basic test with CRUD + import + external changes, with
the possibility to create tests that cover other cases (complete), but
it depends on the resource structure and what can be covered by the
basic one. The pattern used is:

Basic
- Create - without optionals
- Import - without optionals
- Update - set optionals
- Import - with optionals
- Update - unset optionals
- Update - external changes
- Empty config - to ensure no resource is in the state
- Create - with optionals

Complete
- ... more if needed (depending on a resource)

In some of the test files, we ended up merging those two into one, but I
wanted to keep them separate as we need two Create steps, one with and
the second without optional fields set. Alternatively, here are other
approaches I propose:

1) We could use two `resource.Test` functions in one test function,
which wasn't practiced before that much (I think we have one or two such
tests). The advantage of this would be that the objects needed in both
tests would be created only once.

2) Another approach would be to use both models at once. It was also not
done that often before, but it could improve test performance and maybe
test readability.

**Updated:** 3) It turns out there is a Taint option in the Terraform
testing framework that enables the user to have multiple create steps
within one step which eliminates the need to have two test functions
(basic + complete) for basic resources. In the cases of simple resources
we should have only one function (basic) with above operations.

## Next PRs
- See how it applies to a more complex object (e.g. view or warehouse)
Create similar common testing patterns for the following test types:
- Data source tests
- Resources representing connections rather than objects (e.g. grants) 
- Provider tests?