chore: Base for main connection resource use case tests (#4114)
Continuation of
https://github.com/snowflakedb/terraform-provider-snowflake/pull/4113,
but for resources that represent connections rather than a Snowflake
object itself. In this case, the resource and the tests turned out to be
pretty straightforward, as the grant resource(s) don't have many
non-force-new optional fields. The only being is always_apply, which is
thoroughly tested in other tests (thus omitted in the adjusted tests).
Due to the simplicity of the grant resource(s), I suggest only going
with one test covering "complete" use case, which is:
- Create with required fields
- Import
- Update by updating config
- Update by detected external changes

## Next prs
- Adjust other grant tests within the
`resource_grant_privileges_to_account_role_acceptance_test.go` and other
grant test files