chore: Update docker-compose for test output processing (#4148)
Because we planned to use the test output processing like this:
```shell
make run-tests < go run test-output-script > output_file.csv
```

I had to adjust the approach so that it could run on CI. The most
important part is that, because the go CLI is a dependency only
accessible through a Docker container, I had to create another service
for our test output processor. To share the data between the services, I
created a shared volume between them that exists within the containers.

From the CI pov it will be invoked like so (`--silent` in make doesn't
print what underlying command is being run; it's important as otherwise
this command would be put into resulting csv):
```shell
make --silent process-test-output-docker-compose 1> output.csv
```