chore: Base for main resource use case tests follow up 5 (#4132)
Continuation of adjusting resource acceptance tests to the new format.

## Changes
Adjusted resource tests:
- `snowflake_secret_with_generic_string`
- `snowflake_secret_with_basic_authentication`
- `snowflake_secret_with_oauth_authorization_code_grant`
- `snowflake_secret_with_oauth_client_credentials`

During the tests, I encountered an error related to a stale resource
state for each secret resource, which ~I would like to analyze and
resolve in the following PR~ was resolved in this PR and documented in
the migration guide.