chore: Improve the cache logic (#4154)
Before this change, we had only a one-entry cache. Trying to initialize
the provider in another way required setting the
`SF_TF_ACC_TEST_CONFIGURE_CLIENT_ONCE` environment variable, effectively
refreshing the cache. The problem with this approach was that every such
test required refreshing the cache, was slower, and another refresh was
required in the test that followed (if it was using the "normally"
configured provider). In this new approach, we have a cache containing
multiple entries, which should improve the performance and stability of
the tests. There are a few predefined factories (e.g., all view tests
require the warehouse in the session), but it's also possible to define
a new dedicated entry in a test or use the no cache factory. This is the
start but there are a few follow-ups for later, to make it better.

- Simplify the provider initialization caching logic (acceptance tests
and PoCs)
- Remove `SF_TF_ACC_TEST_CONFIGURE_CLIENT_ONCE` and
`testenvs.ConfigureClientOnce` entirely
- Extract a simple `providerInitializationCache`

Needed follow-ups (mostly documented in SNOW-2661409):
- check all the places using TestAccProvider directly
- secondary account can also have a different configuration, so for now
we need to be careful; let's add some hash check for the config or
something else to mitigate
- check which of the `providerFactoryWithoutCache` usages want to really
be without cache and which could utilize a dedicated cache entry
- fix a few tests still requiring the warehouse in session