chore: Regenerate SDK without package problems (#4183)
The main problems solved by this change:
- Compilation problems during the regeneration/clean regeneration -
non-existent interfaces, types - e.g. from ext (SDK generator is no
longer depending on the SDK package which was the main source of issue)
- No separation between example and production SDK objects - causing
problems for both generations (the lists are now separate and only one
is picked based on the build tag from the `generate.go` file)

Changes:
- Remove old generator logic and use the reworked one (only Sequences
moved currently)
- Extract two `0_init.go` files to keep SDK and example definitions
separately. `generate.go` files are adjusted to run with a proper tag.
- Merge `example` and `rework` packages
- Merge sdk_definitions from `example` and `rework` packages
- Clean up Makefile and README
- Remove `rework` Makefile and README
- Regenerate all examples and Sequences
- Clean up SDK generation packages to be more consistent with other
generators, new layout:
  - `sdk`
    - `generator`
      - `defs`
      - `example`
      - `gen`
        - `main`
        - `templates`
        - ... all generator specific files...
      - `README.md`
    - `generate.go`
    - ... all generated and manually added SDK files...

Next PRs:
- move the rest of the SDK definitions (definition files contain enum
definitions - they need to stay in the SDK package; move some manually
added stuff to ext files - as now it won't stop the regeneration)
- clean up essentials/improvements/known issues README sections
- bring back convenience Makefile recipes for a single object removal
and generation (now the generation can be achieved with the object
filtering, but because it's based on the names and not file names now,
the file removal is not that trivial)