chore: Fix data source tests (#4203)
Changes
Adjust data source tests: 
- In most cases the `IN` filtering, if not set, does not lists all
objects `IN ACCOUNT`, but at the `IN DATABASE` level where database is
chosen from the context which depends on the tests execution. Because of
this, some of the tests were adjusted and in the database role case, I
removed one test case, as it was repetitive.
- The warehouse test was failing on one parameter which appears to be
different in different test Snowflake accounts (in different
environments). For now, created a ticket and commented out this part to
find out the reason for this difference.