chore: Fix SDK generation when nested definition is reused (#4250)
Generate SDK correctly when the nested definitions are reused.

The problem lies with the `Field` implementation; the `[]*Field` slice
reuses fields in some cases, causing the setting parent to be incorrect
and resulting in the generation of incorrect paths (which were
previously altered manually).

For now, the implementation has been changed:
- `[]Field` instead of a slice of pointers
- reused nested definitions are wrapped in a function call
- `panic` was added for cases when the parent is being overriden

Later, when we split the definition from the model (as in the rest of
the generators), the field structure should be constructed each time,
allowing for the reuse of nested definitions.