fix: Correct handling of grants to APPLICATION in SHOW GRANTS (#4288)
This commit addresses a bug where the provider failed to correctly
handle the `SHOW GRANTS OF DATABASE ROLE` output when a database role
was granted to an application. The fix ensures that the output is
processed correctly, preventing errors during Terraform plan and apply
operations. Additionally, new acceptance tests have been added to verify
the correct handling of grants to applications.

References:
[#4284](https://github.com/snowflakedb/terraform-provider-snowflake/issues/4284).

---------

Co-authored-by: Jan Cieślak <jan.cieslak@snowflake.com>